requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# OAuth and API clients
httpx==0.25.1
slack-sdk==3.26.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# Task queue
celery==5.3.6
redis==5.0.1

# Security and encryption
cryptography==41.0.5
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
pydantic==2.4.2
pydantic-settings==2.0.3
python-dateutil==2.8.2 